8301283: Util methods for computing text/width height giving up some performance
